requirements.txt
flask
python-pptx
PyPDF2
transformers
torch